Update: Users Management
1. Create
2. Register
3. Change Password
4. Edit

+++ app/Config/Routes.php
@@ -32,21 +32,25 @@ $routes->set404Override();
 
 //Home
 $routes->get('/dashboard', 'Pages::index');
-$routes->get('/', 'Auth::index');
-$routes->get('/logout', 'Auth::delete');
-$routes->post('/login', 'Auth::check');
+$routes->get('/', 'Pages::index');
+
+// For manual auth
+
+// $routes->get('/', 'Auth::index');
+// $routes->post('/login', 'Auth::check');
+// $routes->get('/logout', 'Auth::delete');
 
 //HTTP Spoofing Method
-$routes->delete('item/delete/(:num)', 'Pages::delete/$1');
-$routes->delete('po/delete/(:num)', 'Orders::deletePo/$1');
-$routes->delete('do/delete/(:num)', 'Orders::deleteDo/$1');
+$routes->delete('item/delete/(:num)', 'Pages::delete/$1', ['filter' => 'role:admin']);
+$routes->delete('po/delete/(:num)', 'Orders::deletePo/$1', ['filter' => 'role:admin']);
+$routes->delete('do/delete/(:num)', 'Orders::deleteDo/$1', ['filter' => 'role:admin']);
 
 //Items
 $routes->get('about/', 'Pages::about');
 $routes->get('about/(:any)', 'Pages::detail/$1');
 $routes->get('item/create', 'Pages::create');
-$routes->get('item/edit/(:num)', 'Pages::edit/$1');
-$routes->post('item/update/(:num)', 'Pages::update/$1');
+$routes->get('item/edit/(:num)', 'Pages::edit/$1', ['filter' => 'role:admin']);
+$routes->post('item/update/(:num)', 'Pages::update/$1', ['filter' => 'role:admin']);
 $routes->post('item/save', 'Pages::save');
 
 //Purchase Orders
@@ -54,16 +58,26 @@ $routes->get('po/', 'Orders::index');
 $routes->get('po/create', 'Orders::createPo');
 $routes->post('po/save', 'Orders::savePo');
 $routes->get('detailpo/(:any)', 'Orders::detailPo/$1');
-$routes->get('po/edit/(:num)', 'Orders::editPo/$1');
-$routes->post('po/update/(:num)', 'Orders::updatePo/$1');
+$routes->get('po/edit/(:num)', 'Orders::editPo/$1', ['filter' => 'role:admin']);
+$routes->post('po/update/(:num)', 'Orders::updatePo/$1', ['filter' => 'role:admin']);
 
-//Delivery Orders
+//Mutation Orders
 $routes->get('do/', 'Orders::deliv_order');
 $routes->get('do/create', 'Orders::createDo');
 $routes->post('do/save', 'Orders::saveDo');
 $routes->get('detaildo/(:any)', 'Orders::detailDo/$1');
-$routes->get('do/edit/(:num)', 'Orders::editDo/$1');
-$routes->post('do/update/(:num)', 'Orders::updateDo/$1');
+$routes->get('do/edit/(:num)', 'Orders::editDo/$1', ['filter' => 'role:admin']);
+$routes->post('do/update/(:num)', 'Orders::updateDo/$1', ['filter' => 'role:admin']);
+
+//User
+$routes->get('admin/', 'Users::index', ['filter' => 'role:admin']);
+$routes->get('admin/create', 'Users::createUser', ['filter' => 'role:admin']);
+$routes->post('admin/update/(:num)', 'Users::update/$1', ['filter' => 'role:admin']);
+$routes->get('profile/', 'Users::profile');
+$routes->get('admin/(:num)', 'Users::detail/$1', ['filter' => 'role:admin']);
+$routes->post('changePassword/', 'Users::changePasswordUser');
+
+// Nota
 
 /*
  * --------------------------------------------------------------------
